Refactor auth navigation structure and fix sign up process bug
