+++ navigation/FeedStack.tsx
@@ -1,12 +1,15 @@
 import React from "react";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import FeedComments from "../screens/Feed/FeedComments";
+import ImageSelecter from "../screens/FeedCreation/ImageSelecter";
+import ModifiyFeed from "../screens/Feed/ModifiyFeed";
+import MyClubSelector from "../screens/FeedCreation/MyClubSelector";
 
 const NativeStack = createNativeStackNavigator();
 
 const FeedStack = ({
   route: {
-    params: { feedIndex, feedId, clubId },
+    params: { feedIndex, feedId, clubId, userId, feedData },
   },
   navigation: { navigate },
 }) => {
@@ -27,6 +30,26 @@ const FeedStack = ({
           title: "댓글",
         }}
       />
+
+      <NativeStack.Screen
+        name="ImageSelecter"
+        component={ImageSelecter}
+        initialParams={{ userId, clubId }}
+        options={{
+          title: "",
+          headerShown: true,
+        }}
+      />
+      <NativeStack.Screen
+        name="MyClubSelector"
+        component={MyClubSelector}
+        initialParams={{ userId }}
+        options={{
+          title: "나의 모임",
+          headerShown: true,
+        }}
+      />
+      <NativeStack.Screen name="ModifiyFeed" component={ModifiyFeed} initialParams={{ feedData }} options={{ title: "수정" }} />
     </NativeStack.Navigator>
   );
 };
