+++ navigation/Root.tsx
@@ -3,7 +3,6 @@ import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import Tabs from "./Tabs";
 import ClubCreationStack from "./ClubCreationStack";
 import ClubStack from "./ClubStack";
-import HomeStack from "./HomeStack";
 import ProfileStack from "./ProfileStack";
 import ClubManagementStack from "./ClubManagementStack";
 import FeedStack from "./FeedStack";
@@ -18,7 +17,6 @@ import { DeviceEventEmitter, Platform } from "react-native";
 import { BaseResponse, ErrorResponse, TargetTokenUpdateRequest, UserApi, UserInfoResponse } from "../api";
 import { useMutation, useQuery, useQueryClient } from "react-query";
 import messaging from "@react-native-firebase/messaging";
-import feedSlice from "../redux/slices/feed";
 
 const Nav = createNativeStackNavigator();
 
@@ -158,7 +156,6 @@ const Root = () => {
           <Nav.Screen name="ClubCreationStack" component={ClubCreationStack} />
           <Nav.Screen name="ClubManagementStack" component={ClubManagementStack} />
           <Nav.Screen name="ClubStack" component={ClubStack} />
-          <Nav.Screen name="HomeStack" component={HomeStack} />
           <Nav.Screen name="ProfileStack" component={ProfileStack} />
           <Nav.Screen name="FeedStack" component={FeedStack} />
         </Nav.Navigator>
