+++ screens/SignUp/JoinStepEight.tsx
@@ -1,11 +1,10 @@
 import { NativeStackScreenProps } from "@react-navigation/native-stack";
-import React, { useState, useLayoutEffect } from "react";
-import { Keyboard, TouchableWithoutFeedback, TouchableOpacity } from "react-native";
+import React, { useState, createRef, useLayoutEffect } from "react";
+import { Keyboard, TouchableWithoutFeedback, TouchableOpacity, StatusBar } from "react-native";
 import styled from "styled-components/native";
-import { Entypo } from "@expo/vector-icons";
+import { Entypo, MaterialCommunityIcons } from "@expo/vector-icons";
 import CustomText from "../../components/CustomText";
 import CustomTextInput from "../../components/CustomTextInput";
-import { MaterialCommunityIcons } from "@expo/vector-icons";
 
 const Container = styled.View`
   width: 100%;
@@ -17,7 +16,7 @@ const Container = styled.View`
 
 const Wrap = styled.View`
   width: 100%;
-  padding: 0 20px;
+  padding: 0px 20px;
 `;
 
 const ButtonWrap = styled.View`
@@ -31,7 +30,7 @@ const BorderWrap = styled.View`
 `;
 
 const Border = styled.View`
-  width: 50%;
+  width: 80%;
   height: 2px;
   background-color: #295af5;
 `;
@@ -56,14 +55,14 @@ const Input = styled.TextInput`
   font-size: 18px;
 `;
 
-const Button = styled.TouchableOpacity<{ disabled: boolean }>`
+const Button = styled.TouchableOpacity`
   flex-direction: row;
   justify-content: center;
   align-items: center;
   width: 100%;
   height: 68px;
   padding-bottom: 8px;
-  background-color: ${(props: any) => (props.disabled ? "#D3D3D3" : "#295af5")};
+  background-color: ${(props: any) => (props.disabled ? "#d3d3d3" : "#295AF5")};
 `;
 
 const ButtonTitle = styled(CustomText)`
@@ -71,6 +70,13 @@ const ButtonTitle = styled(CustomText)`
   font-size: 20px;
   line-height: 24px;
   color: #fff;
+`;
+
+const Error = styled.Text`
+  color: #ff6534;
+  font-size: 12px;
+  margin-top: 7px;
+  margin-bottom: 20px;
 `;
 
 const FieldContentView = styled.View`
@@ -89,6 +95,12 @@ const FieldContentOptionLine = styled.View`
   margin-bottom: 15px;
 `;
 
+const ChoiceButton = styled.TouchableOpacity`
+  flex-direction: row;
+  justify-content: flex-start;
+  align-items: center;
+`;
+
 const FieldContentText = styled.Text`
   font-size: 18px;
   margin-right: 10px;
@@ -100,42 +112,36 @@ const SkipText = styled(CustomText)`
   color: #8e8e8e;
 `;
 
-const ChoiceButton = styled.TouchableOpacity`
-  flex-direction: row;
-  justify-content: flex-start;
-  align-items: center;
-`;
-
-const JoinStepFive: React.FC<NativeStackScreenProps<any, "AuthStack">> = ({
+const JoinStepEight: React.FC<NativeStackScreenProps<any, "JoinStepEight">> = ({
   navigation: { navigate, setOptions },
   route: {
-    params: { name, email, password },
+    params: { name, email, password, sex, birth, phone },
   },
 }) => {
-  const [approvalMethod, setApprovalMethod] = useState<number>(0);
+  const [check, setCheck] = useState(1);
 
-  const goToNext = () => {
-    let sex = null;
-    if (approvalMethod === 1) sex = "남성";
-    if (approvalMethod === 2) sex = "여성";
-    navigate("LoginStack", {
-      screen: "JoinStepSix",
+  const validate = () => {
+    navigate("SignUpStack", {
+      screen: "JoinConfirm",
       name,
       email,
       password,
       sex,
+      birth,
+      phone,
+      church: "시광교회",
     });
   };
 
   useLayoutEffect(() => {
     setOptions({
       headerLeft: () => (
-        <TouchableOpacity onPress={() => navigate("LoginStack", { screen: "JoinStepFour", name, email, password })}>
+        <TouchableOpacity onPress={() => navigate("SignUpStack", { screen: "JoinStepSeven", name, email, password, sex, birth, phone })}>
           <Entypo name="chevron-thin-left" size={20} color="black" />
         </TouchableOpacity>
       ),
     });
-  }, [name, email, password]);
+  }, [name, email, password, sex, birth, phone]);
 
   return (
     <TouchableWithoutFeedback
@@ -144,34 +150,27 @@ const JoinStepFive: React.FC<NativeStackScreenProps<any, "AuthStack">> = ({
       }}
     >
       <Container>
+        <StatusBar backgroundColor={"white"} barStyle={"dark-content"} />
         <Wrap>
           <BorderWrap>
             <Border></Border>
           </BorderWrap>
-          <AskText>성별이 어떻게 되시나요?</AskText>
-          <SubText>멤버 관리와, 동명이인 구분을 위함 입니다.</SubText>
+          <AskText>출석중인 교회를 알려주세요.</AskText>
+          <SubText>멤버 관리와, 소모임 소속 기관을 알기 위함 입니다.</SubText>
           <FieldContentView>
             <FieldContentLine>
-              <ChoiceButton onPress={() => setApprovalMethod((prev) => (prev === 1 ? 0 : 1))} activeOpacity={0.5}>
-                <FieldContentText> 남성</FieldContentText>
-                {approvalMethod === 1 ? <MaterialCommunityIcons name="radiobox-marked" size={20} color="#295AF5" /> : <MaterialCommunityIcons name="radiobox-blank" size={20} color="#ABABAB" />}
-              </ChoiceButton>
-            </FieldContentLine>
-            <FieldContentLine>
-              <ChoiceButton onPress={() => setApprovalMethod((prev) => (prev === 2 ? 0 : 2))} activeOpacity={0.5}>
-                <FieldContentText> 여성</FieldContentText>
-                {approvalMethod === 2 ? <MaterialCommunityIcons name="radiobox-marked" size={20} color="#295AF5" /> : <MaterialCommunityIcons name="radiobox-blank" size={20} color="#ABABAB" />}
+              <ChoiceButton onPress={() => setCheck(1)} activeOpacity={0.5}>
+                <FieldContentText>{`시광교회`}</FieldContentText>
+                {check === 1 ? <MaterialCommunityIcons name="radiobox-marked" size={20} color="#295AF5" /> : <MaterialCommunityIcons name="radiobox-blank" size={20} color="#ABABAB" />}
               </ChoiceButton>
             </FieldContentLine>
             <FieldContentOptionLine>
-              <SkipButton onPress={goToNext}>
-                <SkipText>선택하지 않고 넘어가기</SkipText>
-              </SkipButton>
+              <SkipText>{`지금은 시광교회 교인으로만 가입할 수 있습니다.`}</SkipText>
             </FieldContentOptionLine>
           </FieldContentView>
         </Wrap>
         <ButtonWrap>
-          <Button onPress={goToNext}>
+          <Button onPress={validate}>
             <ButtonTitle>다음</ButtonTitle>
           </Button>
         </ButtonWrap>
@@ -180,4 +179,4 @@ const JoinStepFive: React.FC<NativeStackScreenProps<any, "AuthStack">> = ({
   );
 };
 
-export default JoinStepFive;
+export default JoinStepEight;
